+++ example/lib/widgets/shimmer_loader.dart
@@ -17,10 +17,8 @@ class _ShimmerLoaderState extends State<ShimmerLoader> with SingleTickerProvider
   @override
   void initState() {
     super.initState();
-    _controller = AnimationController(
-      vsync: this,
-      duration: const Duration(milliseconds: 1500),
-    )..repeat();
+    _controller = AnimationController(vsync: this, duration: const Duration(milliseconds: 1500))
+      ..repeat();
   }
 
   @override